Implemented quasiquote(unquote, unquote-splicing)
Perfectly done with it. All test cases are passed successfully. This
function works as good as GNU Guile*.

* The test cases are based on revo and guile, in which two it gives
  the same result.

`obj is equivalent to (quasiquote obj), ,obj is equivalent to (unquote
obj), and ,@obj is equivalent to (unquote-splicing obj). The
abbreviated forms are converted into the longer forms by the Scheme
reader (see read).

quasiquote is similar to quote, but it allows parts of the quoted text
to be "unquoted." Within a quasiquote expression, unquote and
unquote-splicing subforms are evaluated, and everything else is
quoted, i.e., left unevaluated. The value of each unquote subform is
inserted into the output in place of the unquote form, while the value
of each unquote-splicing subform is spliced into the surrounding list
or vector structure. unquote and unquote-splicing are valid only
within quasiquote expressions.

quasiquote expressions may be nested, with each quasiquote introducing
a new level of quotation and each unquote or unquote-splicing taking
away a level of quotation. An expression nested within n quasiquote
expressions must be within n unquote or unquote-splicing expressions
to be evaluated.

(See: http://www.scheme.com/tspl4/objects.html#./objects:h1)